issues/14: fixed from using url() to route()

+++ database/migrations/2020_07_07_064150_create_locations_table.php
@@ -17,7 +17,7 @@ class CreateLocationsTable extends Migration
       $table->id();
       $table->string('name', 50);
       $table->tinyInteger('type');
-      $table->integer('parent_id')->nullable()->references('id')->on('locations');
+      $table->unsignedBigInteger('parent_id')->nullable()->references('id')->on('locations');
       $table->timestamp('created_at')->useCurrent();
       $table->timestamp('updated_at')->useCurrent();
     });
